refactor(styles): move the style model into generators
The store defines model, so it lives with the rest of the model code
instead of its own folder. Flat files per the generators layout:
styles.ts, styles-defaults.ts, styles-legacy.ts.
