+++ src/controllers/labels-group-editor.ts
@@ -4,9 +4,9 @@ import { LAYER_TOGGLES } from "@/components/layers-tab";
 import { tip } from "@/components/tooltips";
 import { Controllers } from "@/controllers";
 import { LABEL_TYPES, type LabelGroup, type LabelNameMode, type LabelType } from "@/generators/labels-generator";
+import { styles } from "@/generators/styles";
 import { getLabelsData } from "@/renderers/labels/label-data";
 import { getGroupStyle } from "@/renderers/labels/label-groups";
-import { styles } from "@/styles/styles";
 import { ensureEl } from "@/utils";
 
 function open(): void {
