+++ src/controllers/submap-tool.ts
@@ -1,8 +1,8 @@
 import { destroyDialog } from "@/components/dialog/dialog-helpers";
 import { Layers } from "@/components/layers";
 import { Resample } from "@/generators/resample";
-import { burgGroupFromElement } from "@/styles/legacy";
-import { styles } from "@/styles/styles";
+import { styles } from "@/generators/styles";
+import { burgGroupFromElement } from "@/generators/styles-legacy";
 import { getLatitude, getLongitude } from "@/utils";
 import { ensureEl, minmax, rn } from "../utils";
