+++ src/generators/styles-legacy-default.fixture.json
