+++ src/generators/styles-legacy-serializer.fixture.json
