+++ src/index.html
@@ -5154,7 +5154,7 @@
     <script type="module" src="renderers/index.ts"></script>
     <script type="module" src="controllers/index.ts"></script>
     <script type="module" src="services/index.ts"></script>
-    <script type="module" src="styles/legacy.ts"></script>
+    <script type="module" src="generators/styles-legacy.ts"></script>
     <script defer src="libs/alea.min.js?v1.105.0"></script>
     <script defer src="libs/polylabel.min.js?v1.105.0"></script>
     <script defer src="libs/simplify.js?v1.105.6"></script>
