+++ src/services/io/auto-update.ts
@@ -7,12 +7,12 @@ import { Emblems } from "@/generators/emblems-generator";
 import type { GraphOverrides } from "@/generators/graph-override";
 import type { Label, LabelNameMode } from "@/generators/labels-generator";
 import type { Measurer, MeasurerType } from "@/generators/measurers-generator";
+import { parseStyles, styles } from "@/generators/styles";
+import { isLegacyPreset, labelGroupFromLegacy, presetToLegacy } from "@/generators/styles-legacy";
 import type { Point } from "@/generators/voronoi";
 import { getGroupStyle } from "@/renderers/labels/label-groups";
 import { unfog } from "@/renderers/overlays/fogging";
 import { compareVersions } from "@/services/versioning";
-import { isLegacyPreset, labelGroupFromLegacy, presetToLegacy } from "@/styles/legacy";
-import { parseStyles, styles } from "@/styles/styles";
 import type { ReliefSet } from "@/types/relief";
 import { ensureEl, findEl, P, parseTransform, rand, rn, rw, safeParseJSON, unique } from "@/utils";
 import { parsePathPoints } from "@/utils/pathUtils";
