+++ src/services/io/save.ts
@@ -4,11 +4,10 @@ import { closeDialogs } from "@/components/dialog/dialog-helpers";
 import { Layers } from "@/components/layers";
 import { tip } from "@/components/tooltips";
 import { GraphOverride } from "@/generators/graph-override";
+import { stylesToLegacy } from "@/generators/styles-legacy";
 import { Services } from "@/services";
 import { getUsedFonts } from "@/services/fonts";
-
 import { VERSION } from "@/services/versioning";
-import { stylesToLegacy } from "@/styles/legacy";
 import { ensureEl, getFileName, link, parseError, rn } from "@/utils";
 
 type SaveMethod = "storage" | "machine" | "dropbox";
